Adding changes to display genre, update tracks, and create log in state for users

+++ routes/post_routes/removetrackPR.js
@@ -3,21 +3,36 @@ const removetrackPR = express.Router();
 const connection = require(`../../middleware/db`);
 
 
-removetrackPR.post("/removetrack", (req, res) => {
+removetrackPR.post("/removetrack", (req, res, next) => {
 
     let recordid = req.query.recordid;
     let track_id = req.body.trackid;
+    let input = req.body.edit;
 
-    let removetrack =
-        `DELETE FROM record_track 
+
+    if (input == `remove`) {
+        let removetrack =
+            `DELETE FROM record_track 
         WHERE record_track.track_id = ?;
         
         DELETE FROM track 
         WHERE track.track_id = ?;`;
 
-    connection.query(removetrack, [track_id, track_id], (err, rows) => {
-        if (err) throw err;
-    });
+        connection.query(removetrack, [track_id, track_id], (err, rows) => {
+            if (err) throw err;
+        });
+    } else if (input == `update`) {
+        let trackname = req.body.trackname;
+        let updatetrack =
+            `UPDATE track 
+        SET track_name = ? 
+        WHERE track_id = ?;`;
+
+        connection.query(updatetrack, [trackname, track_id], (err, rows) => {
+            if (err) throw err;
+        });
+    }
+
 
     let readtracks =
         `SELECT track.track_id, track_name, record_name,cover_image, record.record_id FROM track
